release 1 finalized //TODO tests

+++ src/main/java/de/ait_tr/models/Product.java
@@ -33,7 +33,7 @@ public class Product {
         return basicPrice;
     }
 
-    public double getMarkup() {
+    public int getMarkup() {
         return markup;
     }
